feat(ondombeforeinput): make ondombeforeinput preventable and make the event an input event
previously onDOMBeforeInput was run on every extension. Now onDOMBeforeInput can be handled by an
extension by returning false. Also the event is now an InputEvent rather than a generic Event.

BREAKING CHANGE: Change semantics of onDOMBeforeInput so that the event can be handled by an
extension. Strongle type the event type of onDOMBeforeInput as an InputEvent.