feat(middleware): reimplement most editor and slate methods as middleware
previously we had no way of controlling whether the next middleware funciton is run. This change
provides each of the editor methods with a next method which can be used to determine if the next
middleware method is run or not.

BREAKING CHANGE: All methods which were run one at a time such as insertBreak and onChange now
require the user to call next to run the next method. Editor is no longer the first argument for
isInline and isVoid