feat(deps): upgrade dependencies
BREAKING CHANGE: there are now two ways to apply withPlugins. as prePlugins and postPlugins.
prePlugins are applied once to the editor before any extensions are applied and change the editor
for it's entire lifecycle. Think of withReact and withHistory. postPlugins are applied after
extensions are applied and may be modified during the editor lifecycle. For example a custom
decorate method could be used as a postPlugin. postPlugins accept dependencies and can update while
prePlugins never accept dependencies and never update.

fix #2